refactor(line, section) : section에 관련된 클래스를 section 패키지로 이관
