+++ src/main/java/subway/line/LineController.java
@@ -3,6 +3,7 @@ package subway.line;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.section.SectionRequest;
 
 import javax.annotation.Resource;
 import java.util.List;
