+++ src/main/java/subway/section/SectionRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.section;
 
 public class SectionRequest {
     private Long upStationId;
